wallet_consumer/core/serializers.py
from core.models import Account
from rest_framework import serializers


class AccountSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Account
        fields = ['account_id', 'balance']
        