wallet_consumer/requirements.txt
asgiref==3.7.2
confluent-kafka==2.3.0
Django==5.0.3
djangorestframework==3.15.1
mysqlclient==2.2.4
sqlparse==0.4.4
